refactor: config 디렉토리 분리
