+++ src/main/java/yeolJyeongKong/mall/controller/MallController.java
@@ -13,7 +13,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
-import yeolJyeongKong.mall.config.PrincipalDetails;
+import yeolJyeongKong.mall.config.auth.PrincipalDetails;
 import yeolJyeongKong.mall.domain.dto.MallDto;
 import yeolJyeongKong.mall.domain.dto.MallPreviewDto;
 import yeolJyeongKong.mall.service.FavoriteService;
